requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.2.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
pillow==11.2.1
PyMySQL==1.1.1
python-dotenv==1.1.0
pytz==2025.2
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
WTForms==3.2.1
Flask-Mail==0.10.0
Flask-ReCaptcha==0.4.2
